moved through tests to a separate file
